Add warning feature
When registering a function, if the optional warning argument is passed,
that warning will be shown if the version matches.

Note: This introduces a pickel incompatibility (i.e. old pickled
versiondispatched functions cannot be loaded anymore), but since nobody
uses this, there is no point in taking care of it :)